+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 6.3.0
+
 Features:
   - Add `monitoring` recipe with scout gem [#353](https://github.com/platanus/potassium/pull/353)
   - Change CircleCI config to use cache in tests, and remove old cibuild flow [#366](https://github.com/platanus/potassium/pull/366)